move voip mocks to mock file
